Fixing and Creating Abilities
Fixed Slash to where Puddles moves backwards. Created a new Shotgun variety to where its spread is larger, but the bullets are bigger. Created the Slime Punch and the Slime Swipe. The Slime Punch sends out a punch that has high knockback. This pushes puddles forwards. The Slime Swipe sends out a slime claw that swipes the enemies. This has decent knockback and is a fast attack. This also pushes puddles forwards.