Slime Trail and Slime Wall
Slime Trail and Slime Wall abilities are created. Slime Trail leaves slime behind puddles as he walks, currently only dealing damage to enemies. Slime Wall places a wall in front of puddles, allow him to be safe behind it, being able to attack from behind. It only deals damage to enemies once. It was suppose to have health and knockback, but I did not get to that.