Fixed Pickup for 2 secondary and tried to do drop
Pickup works for having 2 secondary abilities. When both abilities are full, and another ability is picked up, the ability in the second slot moves to the first, and the new ability is put into the second slot. This works pretty good. Tried top add the ability dropped but that did not work good. :(